requirements.txt
streamlit
torch
torchvision
opencv-python
numpy
Pillow
segmentation-models-pytorch
scikit-image
